Add comprehensive test data dependencies and weekly execution schedule documentation
- Created `test-data-dependencies-prerequisites.md` to outline test data requirements, dependencies, and prerequisites for API testing across 42 categories.
- Established a dependency graph and phase-wise data requirements for systematic test data management.
- Introduced test data creation scripts, environment variable strategies, and data quality requirements.
- Added `test-execution-weekly-schedule.md` detailing a week-by-week schedule for executing API tests, including daily breakdowns, deliverables, and quality gates.
- Documented risk mitigation strategies and success metrics for effective test execution.

+++ Backend/src/logic/pdf_generator.rs
@@ -35,7 +35,7 @@ impl PdfGenerator {
 
                 // Headers
                 let mut x_pos = 20.0;
-                let col_width = 170.0 / headers.len() as f32;
+                let col_width = 170.0 / headers.len() as f64;
 
                 for head in &headers {
                     current_layer.use_text(head, 12.0, Mm(x_pos), Mm(y_pos), &font_bold);
